backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-unfold==0.37.0
django-import-export==4.1.1
django-taggit==6.1.0
django-tinymce==4.1.0
django-cloudinary-storage==0.3.0
razorpay==1.4.2
gunicorn==23.0.0
whitenoise==6.8.1
python-decouple==3.8
dj-database-url==1.0.0
pillow==10.4.0
celery==5.4.0
redis==5.1.1
python-dotenv==1.0.1
tablib==3.5.0
asgiref==3.9.1
billiard==4.2.2
click==8.3.0
kombu==5.5.4
vine==5.1.0
amqp==5.3.1
prompt-toolkit==3.0.52
python-dateutil==2.9.0.post0
requests==2.32.5
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
certifi==2025.8.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
diff-match-patch==20241021
packaging==25.0
wcwidth==0.2.13
